@@ -2,7 +2,6 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
-import 'package:flutter/widgets.dart';
 
 import 'package:spoiler/models/spoiler_details.dart';
 
@@ -10,8 +9,8 @@ typedef OnReady = Function(SpoilerDetails);
 typedef OnUpdate = Function(SpoilerDetails);
 
 class Spoiler extends StatefulWidget {
-  final Widget header;
-  final Widget child;
+  final Widget? header;
+  final Widget? child;
 
   final bool isOpened;
 
@@ -25,96 +24,122 @@ class Spoiler extends StatefulWidget {
 
   final bool waitFirstCloseAnimationBeforeOpen;
 
-  final OnReady onReadyCallback;
-  final OnUpdate onUpdateCallback;
+  final OnReady? onReadyCallback;
+  final OnUpdate? onUpdateCallback;
 
-  final Key key;
-
-  const Spoiler(
-      {this.key,
-      this.header,
-      this.child,
-      this.isOpened = false,
-      this.leadingArrow = false,
-      this.trailingArrow = false,
-      this.waitFirstCloseAnimationBeforeOpen = false,
-      this.duration,
-      this.onReadyCallback,
-      this.onUpdateCallback,
-      this.openCurve = Curves.easeOutExpo,
-      this.closeCurve = Curves.easeInExpo})
-      : super(key: key);
+  const Spoiler({
+    Key? key,
+    this.header,
+    this.child,
+    this.isOpened = false,
+    this.leadingArrow = false,
+    this.trailingArrow = false,
+    this.waitFirstCloseAnimationBeforeOpen = false,
+    this.duration = const Duration(milliseconds: 400),
+    this.onReadyCallback,
+    this.onUpdateCallback,
+    this.openCurve = Curves.easeOutExpo,
+    this.closeCurve = Curves.easeInExpo,
+  }) : super(key: key);
 
   @override
   SpoilerState createState() => SpoilerState();
 }
 
+@visibleForTesting
 class SpoilerState extends State<Spoiler> with SingleTickerProviderStateMixin {
-  double headerWidth;
-  double headerHeight;
+  late final AnimationController childHeightAnimationController;
+  late Animation<double> childHeightAnimation;
 
-  double childWidth;
-  double childHeight;
+  late final StreamController<bool> isReadyController;
+  late final Stream<bool> isReady;
 
-  AnimationController childHeightAnimationController;
-  Animation<double> childHeightAnimation;
+  late final StreamController<bool> isOpenController;
+  late final Stream<bool> isOpen;
 
-  StreamController<bool> isReadyController = StreamController();
-  Stream<bool> isReady;
+  late bool isOpened;
 
-  StreamController<bool> isOpenController = StreamController();
-  Stream<bool> isOpen;
+  VoidCallback? onUpdateCallbackListener;
 
-  bool isOpened;
+  late final GlobalKey _headerKey;
+  late final GlobalKey _childKey;
 
   @override
   void initState() {
     super.initState();
 
+    _headerKey = GlobalKey(debugLabel: 'spoiler_header');
+    _childKey = GlobalKey(debugLabel: 'spoiler_child');
+
     isOpened = widget.isOpened;
 
+    isReadyController = StreamController<bool>();
     isReady = isReadyController.stream.asBroadcastStream();
 
+    isOpenController = StreamController<bool>();
     isOpen = isOpenController.stream.asBroadcastStream();
 
     childHeightAnimationController = AnimationController(
-        duration: widget.duration != null
-            ? widget.duration
-            : Duration(milliseconds: 400),
-        vsync: this);
+      duration: widget.duration,
+      vsync: this,
+    );
 
     childHeightAnimation = CurvedAnimation(
-        parent: childHeightAnimationController,
-        curve: widget.openCurve,
-        reverseCurve: widget.closeCurve);
+      parent: childHeightAnimationController,
+      curve: widget.openCurve,
+      reverseCurve: widget.closeCurve,
+    );
 
-    SchedulerBinding.instance.addPostFrameCallback((_) async {
-      headerWidth = _headerKey.currentContext.size.width;
-      headerHeight = _headerKey.currentContext.size.height;
+    SchedulerBinding.instance!.addPostFrameCallback((_) async {
+      final headerElement = _headerKey.currentContext;
+      final childElement = _childKey.currentContext;
 
-      childWidth = _childKey.currentContext.size.width;
-      childHeight = _childKey.currentContext.size.height;
+      if (headerElement == null || childElement == null) return;
 
-      childHeightAnimation = Tween(begin: 0.toDouble(), end: childHeight)
-          .animate(childHeightAnimation);
+      final headerElementSize = headerElement.size;
+      final childElementSize = childElement.size;
 
-      if (widget.onUpdateCallback != null) {
-        childHeightAnimation.addListener(() => widget.onUpdateCallback(
+      if (headerElementSize == null || childElementSize == null) return;
+
+      final headerWidth = headerElementSize.width;
+      final headerHeight = headerElementSize.height;
+
+      final childWidth = childElementSize.width;
+      final childHeight = childElementSize.height;
+
+      childHeightAnimation = Tween(
+        begin: 0.0,
+        end: childHeight,
+      ).animate(childHeightAnimation);
+
+      final onUpdateCallback = widget.onUpdateCallback;
+      if (onUpdateCallback != null) {
+        onUpdateCallbackListener = () {
+          onUpdateCallback(
             SpoilerDetails(
-                isOpened: isOpened,
-                headerWidth: headerWidth,
-                headerHeight: headerHeight,
-                childWidth: childWidth,
-                childHeight: childHeightAnimation.value)));
+              isOpened: isOpened,
+              headerWidth: headerWidth,
+              headerHeight: headerHeight,
+              childWidth: childWidth,
+              childHeight: childHeightAnimation.value,
+            ),
+          );
+        };
+
+        childHeightAnimation.addListener(onUpdateCallbackListener!);
       }
 
-      if (widget.onReadyCallback != null) {
-        widget.onReadyCallback(SpoilerDetails(
+      final onReadyCallback = widget.onReadyCallback;
+      if (onReadyCallback != null) {
+        onReadyCallback(
+          SpoilerDetails(
             isOpened: isOpened,
             headerWidth: headerWidth,
             headerHeight: headerHeight,
             childWidth: childWidth,
-            childHeight: childHeight));
+            childHeight: childHeight,
+          ),
+        );
       }
 
       isReadyController.add(true);
@@ -127,7 +152,10 @@ class SpoilerState extends State<Spoiler> with SingleTickerProviderStateMixin {
                   .forward()
                   .orCancel
                   .whenComplete(
-                      () => childHeightAnimationController.reverse().orCancel);
+                  () {
+                    return childHeightAnimationController.reverse().orCancel;
+                  },
+                );
         } else {
           isOpened
               ? await childHeightAnimationController.forward().orCancel
@@ -141,9 +169,15 @@ class SpoilerState extends State<Spoiler> with SingleTickerProviderStateMixin {
 
   @override
   void dispose() {
+    if (onUpdateCallbackListener != null) {
+      childHeightAnimation.removeListener(onUpdateCallbackListener!);
+    }
+
     childHeightAnimationController.dispose();
+
     isOpenController.close();
     isReadyController.close();
+
     super.dispose();
   }
 
@@ -161,91 +195,104 @@ class SpoilerState extends State<Spoiler> with SingleTickerProviderStateMixin {
     }
   }
 
-  final GlobalKey _headerKey = GlobalKey(debugLabel: 'spoiler_header');
-  final GlobalKey _childKey = GlobalKey(debugLabel: 'spoiler_child');
-
   @override
-  Widget build(BuildContext context) => Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          GestureDetector(
-            onTap: () => setState(() {
-              toggle();
-            }),
-            child: Container(
-              key: Key('spoiler_header'),
-              child: Container(
-                key: _headerKey,
-                child: widget.header != null
-                    ? IntrinsicWidth(
-                      child: Row(
-                          children: <Widget>[
-                            widget.leadingArrow
-                                ? (isOpened
-                                    ? Icon(Icons.keyboard_arrow_up)
-                                    : Icon(Icons.keyboard_arrow_down))
-                                : Container(),
-                            widget.header,
-                            widget.trailingArrow
-                                ? (isOpened
-                                    ? Icon(Icons.keyboard_arrow_up)
-                                    : Icon(Icons.keyboard_arrow_down))
-                                : Container()
-                          ],
-                        ),
-                    )
-                    : _buildDefaultHeader(),
+  Widget build(BuildContext context) {
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: <Widget>[
+        GestureDetector(
+          key: const Key('spoiler_header'),
+          onTap: () => setState(() {
+            toggle();
+          }),
+          child: buildHeader(),
+        ),
+        buildChild(),
+      ],
+    );
+  }
+
+  Widget buildHeader() {
+    final header = widget.header;
+
+    return Container(
+      key: _headerKey,
+      child: header != null
+          ? IntrinsicWidth(
+              child: Row(
+                children: <Widget>[
+                  widget.leadingArrow
+                      ? isOpened
+                          ? const Icon(Icons.keyboard_arrow_up)
+                          : const Icon(Icons.keyboard_arrow_down)
+                      : Container(),
+                  header,
+                  widget.trailingArrow
+                      ? isOpened
+                          ? const Icon(Icons.keyboard_arrow_up)
+                          : const Icon(Icons.keyboard_arrow_down)
+                      : Container(),
+                ],
               ),
+            )
+          : buildDefaultHeader(),
+    );
+  }
+
+  Widget buildChild() {
+    return StreamBuilder<bool>(
+      stream: isReady,
+      initialData: false,
+      builder: (context, snapshot) {
+        final isReady = snapshot.data;
+
+        if (isReady == null) return Container();
+
+        if (isReady) {
+          return AnimatedBuilder(
+            animation: childHeightAnimation,
+            builder: (BuildContext context, Widget? child) {
+              return SizedBox(
+                key: isOpened
+                    ? const Key('spoiler_child_opened')
+                    : const Key('spoiler_child_closed'),
+                child: Wrap(
+                  children: <Widget>[
+                    widget.child ?? Container(),
+                  ],
+                ),
+              );
+            },
+          );
+        }
+
+        return Container(
+          key: isOpened
+              ? const Key('spoiler_child_opened')
+              : const Key('spoiler_child_closed'),
+          child: Container(
+            key: _childKey,
+            child: Wrap(
+              children: <Widget>[
+                widget.child ?? Container(),
+              ],
             ),
           ),
-          StreamBuilder<bool>(
-              stream: isReady,
-              initialData: false,
-              builder: (context, snapshot) {
-                if (snapshot.data) {
-                  return AnimatedBuilder(
-                    animation: childHeightAnimation,
-                    builder: (BuildContext context, Widget child) => Container(
-                      key: isOpened
-                          ? Key('spoiler_child_opened')
-                          : Key('spoiler_child_closed'),
-                      height: childHeightAnimation.value > 0
-                          ? childHeightAnimation.value
-                          : 0,
-                      child: Wrap(
-                        children: <Widget>[
-                          widget.child != null ? widget.child : Container()
-                        ],
-                      ),
-                    ),
-                  );
-                } else {
-                  return Container(
-                    key: isOpened
-                        ? Key('spoiler_child_opened')
-                        : Key('spoiler_child_closed'),
-                    child: Container(
-                      key: _childKey,
-                      child: Wrap(
-                        children: <Widget>[
-                          widget.child != null ? widget.child : Container()
-                        ],
-                      ),
-                    ),
-                  );
-                }
-              }),
-        ],
-      );
+        );
+      },
+    );
+  }
 
-  Widget _buildDefaultHeader() => StreamBuilder<bool>(
+  Widget buildDefaultHeader() {
+    return StreamBuilder<bool>(
       stream: isOpen,
       initialData: isOpened,
-      builder: (context, snapshot) => Container(
-          margin: EdgeInsets.all(10),
-          height: 20,
-          width: 20,
-          child: Center(
-              child: Center(child: snapshot.data ? Text('-') : Text('+')))));
+      builder: (context, snapshot) {
+        final isOpened = snapshot.data;
+        if (isOpened == null) return Container();
+        return isOpened ? const Text('-') : const Text('+');
+      },
+    );
+  }
 }
